fix: auditoria de seguridad y calidad completa
SEGURIDAD:
- Eliminar todas las credenciales hardcodeadas de application.properties
- Remover application.properties del tracking de git (ya en .gitignore)
- Proteger Actuator: solo /health publico, resto requiere ADMINISTRADOR
- Corregir enumeracion de usuarios en recuperacion de contrasena
- Corregir IDOR: cambio de contrasena ahora verifica identidad del usuario
- Corregir IDOR: actualizacion de usuario solo via /me (usuario autenticado)
- Corregir IDOR: cambio de estado de reserva verifica propiedad del anfitrion
- Agregar autenticacion al endpoint de MercadoPago
- Remover token de SonarQube hardcodeado del build.gradle
- Remover credenciales de Grafana hardcodeadas del docker-compose
- Remover IP del VPS hardcodeada de prometheus.yml y docker-compose

BUGS:
- Corregir bug en ContrasenaCodigoReinicio: separar creadoEn de expiraEn
- Corregir responder resena: usaba comentario en lugar de mensaje del anfitrion
- Corregir cambiarEstado alojamiento: solo ADMINISTRADOR (no cualquier ANFITRION)
- Corregir repositorio AlojamientoRepository: anfitrionId era String, debe ser Long

CALIDAD:
- Eliminar todos los System.out.println de ReservaServiceImpl y ReservaController
- Eliminar HelloController de prueba
- Cambiar Float a Double en precios y coordenadas geograficas
- Eliminar dependencias de email redundantes (4 libs -> 1 spring-boot-starter-mail)
- Actualizar imagen Docker a eclipse-temurin:17-jre-jammy (menos vulnerabilidades)
- Mejorar GlobalExceptionHandler: no exponer detalles internos al cliente
- Mejorar MercadoPagoService: back URLs configurables via variables de entorno
- Agregar .env.example con todas las variables requeridas documentadas
- DataInitializer: credenciales admin via variables de entorno ADMIN_EMAIL/ADMIN_PASSWORD

+++ src/main/java/poo/uniquindio/edu/co/Homa/dto/request/AlojamientoRequest.java
@@ -34,16 +34,16 @@ public class AlojamientoRequest {
     @NotNull(message = "La latitud es obligatoria")
     @DecimalMin(value = "-90.0", message = "La latitud debe estar entre -90 y 90")
     @DecimalMax(value = "90.0", message = "La latitud debe estar entre -90 y 90")
-    private Float latitud;
+    private Double latitud;
 
     @NotNull(message = "La longitud es obligatoria")
     @DecimalMin(value = "-180.0", message = "La longitud debe estar entre -180 y 180")
     @DecimalMax(value = "180.0", message = "La longitud debe estar entre -180 y 180")
-    private Float longitud;
+    private Double longitud;
 
     @NotNull(message = "El precio por noche es obligatorio")
     @DecimalMin(value = "0.01", message = "El precio debe ser mayor a 0")
-    private Float precioPorNoche;
+    private Double precioPorNoche;
 
     @NotNull(message = "El máximo de huéspedes es obligatorio")
     @Min(value = 1, message = "Debe permitir al menos 1 huésped")
